# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: COROC API
  version: 0.2.0
  summary: Gestión y cobro de préstamos personales
  description: |
    Contrato de la API de COROC (§24.1-3 del prompt maestro). Versión 0.2.0: operaciones de la Fase 1 implementadas y verificadas
    contra este contrato en cada prueba (el servidor valida las peticiones con él y las pruebas validan las respuestas).

    **Convenciones**
    - Dinero: enteros en la unidad mínima de la moneda (`COP` = peso, `BRL`/`USD` = centavo). Nunca decimales binarios.
    - Tasas: texto decimal (`"0.02"` equivale a dos por ciento).
    - Fechas civiles `YYYY-MM-DD` en la zona horaria de la empresa; instantes en ISO 8601 con zona.
    - Errores con `application/problem+json` (RFC 9457).
    - Toda petición autenticada opera dentro de la empresa del token; la base de datos aplica Row Level Security.
    - Idempotencia: las operaciones que crean movimientos contables aceptan el encabezado `Idempotency-Key`.
    - Concurrencia optimista: los recursos editables exponen `version`; enviar `If-Match` con ese valor.
    - Tiempo real: `GET /events` (Server-Sent Events) emite `payment.posted`, `intake.updated`, `message.updated`, `dashboard.changed`.
  contact:
    name: COROC
  license:
    name: Propietaria
    identifier: LicenseRef-COROC-Proprietary
servers:
  - url: https://api.coroc.app/v1
    description: Producción
  - url: http://localhost:3000/v1
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Ingreso con usuario y contraseña
      description: Bloqueo progresivo tras 5 intentos fallidos (15 min). Si el usuario tiene TOTP, responde `mfa_required`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant, username, password, deviceId]
              properties:
                tenant: { type: string, description: Identificador corto de la empresa, maxLength: 40 }
                username: { type: string, maxLength: 32 }
                password: { type: string, format: password, maxLength: 128 }
                deviceId: { type: string, minLength: 8, maxLength: 120 }
                deviceName: { type: string, maxLength: 120 }
      responses:
        '200':
          description: Sesión iniciada o segundo factor requerido
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Session'
                  - type: object
                    required: [mfa_required, challengeId]
                    properties:
                      mfa_required: { type: boolean, const: true }
                      challengeId: { type: string }
        '401': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /auth/mfa:
    post:
      tags: [auth]
      operationId: verifyMfa
      summary: Verificar código TOTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, code]
              properties:
                challengeId: { type: string }
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refresh
      summary: Renovar tokens (rotación del token de renovación)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken, deviceId]
              properties:
                refreshToken: { type: string }
                deviceId: { type: string }
      responses:
        '200':
          description: Nuevos tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Cerrar la sesión actual
      responses:
        '204': { description: Sesión cerrada }
        '4XX': { $ref: '#/components/responses/Problem' }
  /auth/password/forgot:
    post:
      tags: [auth]
      operationId: forgotPassword
      summary: Enviar enlace de recuperación de un solo uso (30 min)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant, username]
              properties:
                tenant: { type: string }
                username: { type: string }
      responses:
        '202': { description: 'Si el usuario existe, se envió el correo' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /auth/password/reset:
    post:
      tags: [auth]
      operationId: resetPassword
      summary: Crear una contraseña nueva con el enlace de recuperación (un solo uso, 30 min)
      description: Cierra todas las sesiones del usuario.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, format: password, minLength: 12, maxLength: 128 }
      responses:
        '204': { description: Contraseña cambiada }
        '4XX': { $ref: '#/components/responses/Problem' }
  /me:
    get:
      tags: [auth]
      operationId: me
      summary: Usuario actual y permisos
      responses:
        '200':
          description: Usuario
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '4XX': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [auth]
      operationId: updateMe
      summary: Preferencias del usuario (idioma §6, apariencia §5.4, bloqueo por inactividad §7.1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                lang: { $ref: '#/components/schemas/Lang' }
                theme: { type: string, enum: [system, light, dark] }
                name: { type: string, minLength: 1, maxLength: 120 }
                autoLockMinutes: { type: integer, minimum: 1, maximum: 60 }
                email: { type: [string, 'null'], format: email }
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /me/password:
    post:
      tags: [auth]
      operationId: changePassword
      summary: Cambiar la contraseña (cierra las demás sesiones)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password, maxLength: 128 }
                newPassword: { type: string, format: password, minLength: 12, maxLength: 128 }
      responses:
        '204': { description: Contraseña cambiada }
        '4XX': { $ref: '#/components/responses/Problem' }
  /me/mfa/enroll:
    post:
      tags: [auth]
      operationId: enrollMfa
      summary: Iniciar la verificación en dos pasos (TOTP); obligatoria para el Propietario
      responses:
        '200':
          description: Secreto para la app autenticadora
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MfaEnrollment' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /me/mfa/confirm:
    post:
      tags: [auth]
      operationId: confirmMfa
      summary: Confirmar con el primer código; devuelve tokens con el segundo factor activo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: Sesión con segundo factor
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /me/mfa:
    delete:
      tags: [auth]
      operationId: disableMfa
      summary: Desactivar el segundo factor (no disponible para el Propietario)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '204': { description: Desactivado }
        '4XX': { $ref: '#/components/responses/Problem' }
  /me/sessions:
    get:
      tags: [auth]
      operationId: mySessions
      summary: Sesiones abiertas del usuario por dispositivo
      responses:
        '200':
          description: Sesiones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SessionInfo' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /me/sessions/{id}:
    delete:
      tags: [auth]
      operationId: revokeMySession
      summary: Cerrar una sesión propia en otro dispositivo
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Cerrada }
        '4XX': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acceso de 15 minutos ligado al dispositivo.
    webhookSecret:
      type: apiKey
      in: header
      name: X-COROC-Webhook-Secret
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: Código estable, p. ej. RATE_CAP_EXCEEDED, DUPLICATE_RECEIPT, CONTACT_RULE_BLOCKED, VERSION_CONFLICT
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    Currency: { type: string, enum: [COP, BRL, USD] }
    Lang: { type: string, enum: [es, pt-BR, en] }
    Role: { type: string, enum: [owner, admin, collector, auditor] }
...
    Session:
      type: object
      required: [accessToken, refreshToken, expiresIn, user, company, mfaEnrollmentRequired]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string, description: 'Opaco y rotativo; ligado al dispositivo' }
        expiresIn: { type: integer, example: 900 }
        mfaEnrollmentRequired: { type: boolean, description: 'El Propietario debe activar el segundo factor antes de continuar' }
        user: { $ref: '#/components/schemas/User' }
        company:
          type: object
          properties:
            id: { type: string, format: uuid }
            slug: { type: string }
            name: { type: string }
            currency: { $ref: '#/components/schemas/Currency' }
            country: { type: string, enum: [CO, BR, US] }
            timezone: { type: string }
            lang: { $ref: '#/components/schemas/Lang' }
    MfaEnrollment:
      type: object
      required: [secret, otpauthUri]
      properties:
        secret: { type: string, description: Secreto en Base32 para escribirlo a mano }
        otpauthUri: { type: string, description: 'URI otpauth:// para mostrar como código QR' }
    SessionInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        deviceId: { type: string }
        deviceName: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time }
        current: { type: boolean }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        name: { type: string }
        email: { type: [string, 'null'] }
        role: { $ref: '#/components/schemas/Role' }
        lang: { $ref: '#/components/schemas/Lang' }
        theme: { type: string, enum: [system, light, dark] }
        active: { type: boolean }
        mfaEnabled: { type: boolean }
        autoLockMinutes: { type: integer }
        lastLoginAt: { type: [string, 'null'], format: date-time }
        permissions: { type: array, items: { type: string } }
...